oai-mep/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.*
clickclick==20.10.2
connexion==2.13.0
Flask==2.0.3
Flask-PyMongo==2.3.0
idna==3.3
importlib-metadata>=3
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.0.0
MarkupSafe==2.0.1
packaging==21.3
pymongo==4.0.2
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.6.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
swagger-spec-validator==2.7.4
swagger-ui-bundle==0.0.9
urllib3==1.26.9
Werkzeug==2.0.0
zipp==3.6.0
